% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationMatrixGroup.R
\name{AnnotationMatrixGroup}
\alias{AnnotationMatrixGroup}
\title{Single-cell Annotation Matrix Group}
\description{
Class for representing a TileDB group containing one or more
\code{\link{AnnotationMatrix}} arrays that share a common dimension name.
}
\section{Super class}{
\code{\link[tiledbsc:TileDBGroup]{tiledbsc::TileDBGroup}} -> \code{AnnotationMatrixGroup}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{arrays}}{Named list of \code{\link{AnnotationMatrix}} arrays in the group}

\item{\code{dimension_name}}{Name of the dimension that is shared by all arrays
in the group.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnnotationMatrixGroup$new()}}
\item \href{#method-add_annotation_matrix}{\code{AnnotationMatrixGroup$add_annotation_matrix()}}
\item \href{#method-clone}{\code{AnnotationMatrixGroup$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_metadata}{\code{tiledbsc::TileDBGroup$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_metadata}{\code{tiledbsc::TileDBGroup$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_object_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_object_uris}{\code{tiledbsc::TileDBGroup$list_object_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_objects">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_objects}{\code{tiledbsc::TileDBGroup$list_objects()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new AnnotationMatrix group object. The existing
array group is opened at the specified \code{uri} if one is present,
otherwise a new array group is created. The \code{arrays} field is populated
with \code{AnnotationMatrix} objects for each \code{AnnotationMatrix} object
discovered within the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationMatrixGroup$new(uri, dimension_name, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB group}

\item{\code{dimension_name}}{Name of the dimension shared by all arrays within
the group (typically \code{obs_id} or \code{var_id}).}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_annotation_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-add_annotation_matrix}{}}}
\subsection{Method \code{add_annotation_matrix()}}{
Add a new \code{\link{AnnotationMatrix}} array to the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationMatrixGroup$add_annotation_matrix(data, name, metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \code{\link{matrix}} of annotation data to ingest. The \code{matrix} rows
must be aligned to the \code{\link{SCGroup}} dimension indicated by the group's
\code{dimension_name}.}

\item{\code{name}}{Name of the new variable annotation matrix.}

\item{\code{metadata}}{Named list of metadata to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationMatrixGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
