% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPairwiseMatrixGroup.R
\name{AnnotationPairwiseMatrixGroup}
\alias{AnnotationPairwiseMatrixGroup}
\title{Single-cell Annotation Matrix Group}
\description{
Class for representing a TileDB group containing one or more
\code{\link{AnnotationPairwiseMatrix}} arrays that share a common dimension name.
}
\section{Super classes}{
\code{\link[tiledbsc:TileDBGroup]{tiledbsc::TileDBGroup}} -> \code{\link[tiledbsc:AnnotationMatrixGroup]{tiledbsc::AnnotationMatrixGroup}} -> \code{AnnotationPairwiseMatrixGroup}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_matrix}{\code{AnnotationPairwiseMatrixGroup$add_matrix()}}
\item \href{#method-add_seurat_graph}{\code{AnnotationPairwiseMatrixGroup$add_seurat_graph()}}
\item \href{#method-clone}{\code{AnnotationPairwiseMatrixGroup$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_metadata}{\code{tiledbsc::TileDBGroup$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_arrays">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_arrays}{\code{tiledbsc::TileDBGroup$get_arrays()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_metadata}{\code{tiledbsc::TileDBGroup$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="initialize">}\href{../../tiledbsc/html/TileDBGroup.html#method-initialize}{\code{tiledbsc::TileDBGroup$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_object_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_object_uris}{\code{tiledbsc::TileDBGroup$list_object_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_objects">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_objects}{\code{tiledbsc::TileDBGroup$list_objects()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="AnnotationMatrixGroup" data-id="add_annotation_matrix">}\href{../../tiledbsc/html/AnnotationMatrixGroup.html#method-add_annotation_matrix}{\code{tiledbsc::AnnotationMatrixGroup$add_annotation_matrix()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-add_matrix}{}}}
\subsection{Method \code{add_matrix()}}{
Add a new \code{\link{AnnotationPairwiseMatrix}} array to the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$add_matrix(data, name, metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \code{\link{matrix}} of annotation data to ingest. The \code{matrix} rows
must be aligned to the \code{\link{SCGroup}} dimension indicated by the group's
\code{dimension_name}.}

\item{\code{name}}{Name of the new pairwise annotation matrix.}

\item{\code{metadata}}{Named list of metadata to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_seurat_graph"></a>}}
\if{latex}{\out{\hypertarget{method-add_seurat_graph}{}}}
\subsection{Method \code{add_seurat_graph()}}{
Convert a \code{\link[SeuratObject:Graph-class]{SeuratObject::Graph}} object to
\code{\link{AnnotationPairwiseMatrix}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$add_seurat_graph(object, technique)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[SeuratObject:Graph-class]{SeuratObject::Graph}} object.}

\item{\code{technique}}{Name of the technique used to generate the graph
(typically, \code{nn} or \code{snn}).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\subsection{On-Disk Format}{

Arrays are named \verb{graph_<technique>}.
}

\subsection{Metadata}{
\itemize{
\item \code{assay_used}: Name of the assay used to generate the graph.
\item \code{graph_technique}: Name of the technique used to generate the graph.
used.
}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
