% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayMatrixGroup.R
\name{AssayMatrixGroup}
\alias{AssayMatrixGroup}
\title{Single-cell Assay Matrix Group}
\description{
Class for representing a TileDB group containing one or more
\code{\link{AssayMatrix}} arrays that share the same dimensions.
}
\section{Super class}{
\code{\link[tiledbsc:TileDBGroup]{tiledbsc::TileDBGroup}} -> \code{AssayMatrixGroup}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_assay_matrix}{\code{AssayMatrixGroup$add_assay_matrix()}}
\item \href{#method-clone}{\code{AssayMatrixGroup$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_metadata}{\code{tiledbsc::TileDBGroup$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_arrays">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_arrays}{\code{tiledbsc::TileDBGroup$get_arrays()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_metadata}{\code{tiledbsc::TileDBGroup$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="initialize">}\href{../../tiledbsc/html/TileDBGroup.html#method-initialize}{\code{tiledbsc::TileDBGroup$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_object_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_object_uris}{\code{tiledbsc::TileDBGroup$list_object_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_objects">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_objects}{\code{tiledbsc::TileDBGroup$list_objects()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_assay_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-add_assay_matrix}{}}}
\subsection{Method \code{add_assay_matrix()}}{
Add a new \code{\link{AssayMatrix}} array to the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayMatrixGroup$add_assay_matrix(
  data,
  name,
  value_col = "value",
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \code{\link{matrix}} of annotation data to ingest. The \code{matrix} rows
must be aligned to the \code{\link{SCGroup}} dimension indicated by the group's
\code{dimension_name}.}

\item{\code{name}}{Name for the new array, nested with the group's URI.}

\item{\code{value_col}}{Name to use for the TileDB array's attribute that will
contain the matrix values.}

\item{\code{metadata}}{Named list of metadata to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayMatrixGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
