% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledbImageArray.R
\name{create_image_array}
\alias{create_image_array}
\title{Create a TileDB array suitable for storing pixel data.}
\usage{
create_image_array(array_uri, width, height)
}
\arguments{
\item{array_uri}{Path to array to create}

\item{width}{Number of columns in array domain}

\item{height}{Number of rows in array domain}
}
\description{
Create a TileDB array suitable for storing pixel data.
}
