% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h510x.R
\name{H510xToTiledb}
\alias{H510xToTiledb}
\title{Convert 10X H5 file to TileDB}
\usage{
H510xToTiledb(
  source,
  dest = NULL,
  overwrite = FALSE,
  verbose = TRUE,
  attr_type = NULL,
  cell_order = "HILBERT",
  tile_order = "ROW_MAJOR",
  capacity = 10000
)
}
\arguments{
\item{source}{Source HDF5 file path}

\item{dest}{Destination URI for the TileDB array.}

\item{overwrite}{Overwrite existing array at \code{dest} URI (default: \code{FALSE}).}

\item{verbose}{Show progress updates.}

\item{attr_type}{Should the data attribute type be 'integer' or 'double'. By
default this is inferred.}

\item{cell_order, tile_order}{Configure the TileDB array's global cell
ordering by specifying the tile (default: \code{"HILBERT"}) and cell (default:
\code{"ROW_MAJOR"}) ordering. See  \href{https://docs.tiledb.com/main/basic-concepts/terminology}{the docs} for more
information.}

\item{capacity}{Capacity of sparse fragments (default: 10000)}
}
\description{
Create a TileDB sparse array containing single cell data from a 10X-formatted
HDF5 file.
}
