% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledbImageArray.R
\name{ImagePositionstoTileDB}
\alias{ImagePositionstoTileDB}
\title{Create a TileDB array for Visium Image Positions}
\usage{
ImagePositionstoTileDB(file_path, uri, verbose = TRUE)
}
\arguments{
\item{file_path}{Path to CSV file containing 10X visium positions}
}
\description{
Create a TileDB array for Visium Image Positions
}
