% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-apply.R
\name{partition_apply}
\alias{partition_apply}
\title{Partition-wise Processing}
\usage{
partition_apply(
  x,
  fun,
  partition_dim,
  partition_count,
  partition_index = NULL,
  layer = NULL,
  combine = NULL,
  verbose = TRUE,
  tiledbcloud_namespace = NULL
)
}
\arguments{
\item{x}{A \code{\link{SOMA}} object.}

\item{fun}{The callback function to apply to each partition of data.}

\item{partition_dim}{Should the data be partitioned by \code{"obs"} or \code{"var"}?}

\item{partition_count}{The number of partitions to create.}

\item{partition_index}{Optional partition index. If provided only the indexed
partition is processed—useful distributing computations across multiple
nodes/workers.}

\item{layer}{The name of the \code{X} layer to apply the function to (defaults to
the first layer).}

\item{combine}{How should the partitioned results be combined? Can be one of:
\itemize{
\item \code{NULL}: (default) return the list of partitioned results without combining
\item \code{"c"}: concatenate results into a vector
\item \code{"rbind"}: combine matrices by rows
\item \code{"cbind"}: combine matrices by columns
}}

\item{verbose}{Whether to print progress messages.}
}
\description{
Apply a function to user specified partitions of assay data in a SOMA X
layer.
}
\details{
Reads all values from either the \code{obs} or \code{var} dimension the \code{SOMA} and
divides them into \code{partition_count} partitions. If a query was previously
applied to the \code{SOMA}, then that subset of dimensions is partitioned instead.

The function \code{fun} must accept a \code{\link{dgTMatrix}}, which is how each partition
is read into memory.

*Note setting \code{combine} to \code{"rbind"} or \code{"cbind"} uses the internal
helper functions \code{rbind_matrix()}/\code{cbind_matrix()}, respectively, which
pad matrices to ensure non-binding dimensions have the same length and,
importantly, the same order. See \code{\link{bind_matrix}} for more information.
}
