% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb10x.R
\name{Read10x_tiledb}
\alias{Read10x_tiledb}
\title{Read 10x TileDB Array}
\usage{
Read10x_tiledb(uri, verbose = TRUE)
}
\arguments{
\item{uri}{URI for the TileDB array.}

\item{verbose}{Show progress updates.}
}
\value{
Returns a \code{\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}} with feature and barcode
dimensions.
}
\description{
Read 10X CellRanger count data stored in TileDB.
}
