% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGroup.R
\name{SCGroup}
\alias{SCGroup}
\title{Single-cell Group}
\description{
Single-cell Group

Single-cell Group
}
\details{
Class for representing a group of TileDB arrays that consitute an \code{sc_group}, which includes:
\itemize{
\item \code{X} (\code{\link{SCGroup_X}}): a labeled 2D sparse array
\item \code{obs} (\code{\link{SCGroup_Annotation}}): 1D labeled array containing column labels for \code{X}
\item \code{var} (\code{\link{SCGroup_Annotation}}): 1D labeled array containing row labels for \code{X}
}
}
\section{Super class}{
\code{\link[tiledbsc:TiledbBase]{tiledbsc::TiledbBase}} -> \code{SCGroup}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SCGroup$new()}}
\item \href{#method-from_seurat}{\code{SCGroup$from_seurat()}}
\item \href{#method-to_seurat}{\code{SCGroup$to_seurat()}}
\item \href{#method-clone}{\code{SCGroup$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="add_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-add_metadata}{\code{tiledbsc::TiledbBase$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="get_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-get_metadata}{\code{tiledbsc::TiledbBase$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="tiledb_array">}\href{../../tiledbsc/html/TiledbBase.html#method-tiledb_array}{\code{tiledbsc::TiledbBase$tiledb_array()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCGoup object. The existing array group is
opened at the specified \code{group_uri} if one is present, otherwise a new
array group is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$new(array_uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Print status messages}

\item{\code{array}}{URI of the TileDB group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_seurat"></a>}}
\if{latex}{\out{\hypertarget{method-from_seurat}{}}}
\subsection{Method \code{from_seurat()}}{
Convert a Seurat object to a TileDB-backed sc_group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$from_seurat(object, assay = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[Seurat:Seurat-package]{Seurat::Seurat}} object.}

\item{\code{assay}}{Name of the assay to retrieve from the Seurat object. By
default the active assay is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat}{}}}
\subsection{Method \code{to_seurat()}}{
Convert to a Seurat object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$to_seurat(
  project = "SeuratProject",
  assay = "Spatial",
  slice = "slice1",
  filter.matrix = TRUE,
  to.upper = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
