% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGroup_Annotation.R
\name{SCGroup_Annotation}
\alias{SCGroup_Annotation}
\title{Single-cell Group Annotation}
\description{
Single-cell Group Annotation

Single-cell Group Annotation
}
\details{
Base class for representing annotations for the \code{\link{SCGroup_X}} data matrix's
observations and variables. Annotation data is stored as a \code{data.frame} with
the number of rows equal to the length of the corresponding \code{X} dimension,
and rownames matching the values of the corresponding \code{X} dimension's labels.
}
\section{Super class}{
\code{\link[tiledbsc:TiledbBase]{tiledbsc::TiledbBase}} -> \code{SCGroup_Annotation}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array_uri}}{URI of the TileDB array}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SCGroup_Annotation$new()}}
\item \href{#method-from_dataframe}{\code{SCGroup_Annotation$from_dataframe()}}
\item \href{#method-to_dataframe}{\code{SCGroup_Annotation$to_dataframe()}}
\item \href{#method-clone}{\code{SCGroup_Annotation$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="add_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-add_metadata}{\code{tiledbsc::TiledbBase$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="dimensions">}\href{../../tiledbsc/html/TiledbBase.html#method-dimensions}{\code{tiledbsc::TiledbBase$dimensions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="dimnames">}\href{../../tiledbsc/html/TiledbBase.html#method-dimnames}{\code{tiledbsc::TiledbBase$dimnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="get_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-get_metadata}{\code{tiledbsc::TiledbBase$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="schema">}\href{../../tiledbsc/html/TiledbBase.html#method-schema}{\code{tiledbsc::TiledbBase$schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="tiledb_array">}\href{../../tiledbsc/html/TiledbBase.html#method-tiledb_array}{\code{tiledbsc::TiledbBase$tiledb_array()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCObject_Annotation object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_Annotation$new(array_uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array_uri}}{URI of the TileDB array}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-from_dataframe}{}}}
\subsection{Method \code{from_dataframe()}}{
Ingest annotation data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_Annotation$from_dataframe(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a \code{\link{data.frame}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Retrieve the annotation data from TileDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_Annotation$to_dataframe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link{data.frame}} with row names containing values from the index
dimension
data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_Annotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
