% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGroup_X.R
\name{SCGroup_X}
\alias{SCGroup_X}
\title{Single-cell Group: X Matrix}
\description{
Single-cell Group: X Matrix

Single-cell Group: X Matrix
}
\examples{


}
\section{Super class}{
\code{\link[tiledbsc:TiledbBase]{tiledbsc::TiledbBase}} -> \code{SCGroup_X}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SCGroup_X$new()}}
\item \href{#method-from_matrix}{\code{SCGroup_X$from_matrix()}}
\item \href{#method-to_dataframe}{\code{SCGroup_X$to_dataframe()}}
\item \href{#method-to_matrix}{\code{SCGroup_X$to_matrix()}}
\item \href{#method-clone}{\code{SCGroup_X$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="add_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-add_metadata}{\code{tiledbsc::TiledbBase$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="get_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-get_metadata}{\code{tiledbsc::TiledbBase$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="tiledb_array">}\href{../../tiledbsc/html/TiledbBase.html#method-tiledb_array}{\code{tiledbsc::TiledbBase$tiledb_array()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCObject_X object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$new(array_uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-from_matrix}{}}}
\subsection{Method \code{from_matrix()}}{
Ingest assay data from a sparse matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$from_matrix(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a \code{\link[Matrix:dgCMatrix]{Matrix::dgCMatrix}} with string dimensions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Retrieve the assay data from TileDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$to_dataframe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix]{Matrix::dgTMatrix}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-to_matrix}{}}}
\subsection{Method \code{to_matrix()}}{
Retrieve the assay data from TileDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$to_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix]{Matrix::dgTMatrix}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
