% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMACollection.R
\name{SOMACollection}
\alias{SOMACollection}
\title{SOMA Collection}
\description{
Class for representing a \code{SOMACollection}, which may contain of one or more
\code{\link{SOMA}}s.
}
\section{Super class}{
\code{\link[tiledbsc:TileDBGroup]{tiledbsc::TileDBGroup}} -> \code{SOMACollection}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{misc}}{Named list of miscellaneous objects.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{somas}}{Retrieve \code{\link{SOMA}} members.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SOMACollection$new()}}
\item \href{#method-from_seurat}{\code{SOMACollection$from_seurat()}}
\item \href{#method-to_seurat}{\code{SOMACollection$to_seurat()}}
\item \href{#method-soma_uris}{\code{SOMACollection$soma_uris()}}
\item \href{#method-clone}{\code{SOMACollection$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_member">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_member}{\code{tiledbsc::TileDBGroup$add_member()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_metadata}{\code{tiledbsc::TileDBGroup$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="class">}\href{../../tiledbsc/html/TileDBGroup.html#method-class}{\code{tiledbsc::TileDBGroup$class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="count_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-count_members}{\code{tiledbsc::TileDBGroup$count_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_member">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_member}{\code{tiledbsc::TileDBGroup$get_member()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_members}{\code{tiledbsc::TileDBGroup$get_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_metadata}{\code{tiledbsc::TileDBGroup$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="group_exists">}\href{../../tiledbsc/html/TileDBGroup.html#method-group_exists}{\code{tiledbsc::TileDBGroup$group_exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_member_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_member_uris}{\code{tiledbsc::TileDBGroup$list_member_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_members}{\code{tiledbsc::TileDBGroup$list_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_object_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_object_uris}{\code{tiledbsc::TileDBGroup$list_object_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_objects">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_objects}{\code{tiledbsc::TileDBGroup$list_objects()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="print">}\href{../../tiledbsc/html/TileDBGroup.html#method-print}{\code{tiledbsc::TileDBGroup$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="tiledb_group">}\href{../../tiledbsc/html/TileDBGroup.html#method-tiledb_group}{\code{tiledbsc::TileDBGroup$tiledb_group()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SOMACollection}. The existing array group is
opened at the specified array \code{uri} if one is present, otherwise a new
array group is created. The \code{members} field is populated with
\code{SOMA} objects for each URI passed explicitly to \code{soma_uris}, as
well \code{SOMA} objects discovered within the \code{SOMACollection} object's
TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollection$new(uri, verbose = TRUE, config = NULL, ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB group}

\item{\code{verbose}}{Print status messages}

\item{\code{config}}{optional configuration}

\item{\code{ctx}}{optional tiledb context}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_seurat"></a>}}
\if{latex}{\out{\hypertarget{method-from_seurat}{}}}
\subsection{Method \code{from_seurat()}}{
Convert a Seurat object to a TileDB-backed \code{SOMACollection}.
\subsection{Assays}{

Each \verb{[SeuratObject::Assay}] is converted to a \code{\link{SOMA}} and written to
a nested TileDB group with a URI of \verb{./soma_<assay>} where \verb{<assay>}
is the name of the Seurat assay.
}

\subsection{Identities}{

Cell identities in the \code{\link[SeuratObject:Seurat-class]{SeuratObject::Seurat}} are maintained by
creating an \code{active_ident} attribute in \code{obs} that stores the factor
levels as a character vector.
}

\subsection{Dimensionality Reductions}{

Dimensionality reduction results are stored as \code{obsm} and \code{varm} arrays
within an \code{SOMA}. The \code{\link[SeuratObject:DimReduc-class]{SeuratObject::DimReduc}} object's \code{key} slot is
used to determine which \code{SOMA} to store the results in. The array names
are \verb{(obsm|varm)_dimreduction_<name>}, where \verb{<name>} is the name of the
dimensionality reduction method (e.g., \code{"pca"}).
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollection$from_seurat(object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[SeuratObject:Seurat-class]{SeuratObject::Seurat}} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat}{}}}
\subsection{Method \code{to_seurat()}}{
Convert to a \link[SeuratObject:Seurat-class]{SeuratObject::Seurat} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollection$to_seurat(project = "SeuratProject")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{\code{\link[SeuratObject:Project]{SeuratObject::Project}} name for the \code{Seurat} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-soma_uris"></a>}}
\if{latex}{\out{\hypertarget{method-soma_uris}{}}}
\subsection{Method \code{soma_uris()}}{
List the \code{\link{SOMA}} URIs in the collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollection$soma_uris()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of URIs for each \code{\link{SOMA}} in the collection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SOMACollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
