% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-apply.R
\name{tiledb_cloud_partition_apply}
\alias{tiledb_cloud_partition_apply}
\title{Build and execute task graph to slice each partition from the array in
parallel and apply the supplied function.}
\usage{
tiledb_cloud_partition_apply(
  uri,
  fun,
  dimname,
  partitions,
  combine_fun,
  namespace,
  verbose
)
}
\arguments{
\item{uri}{URI for an AssayMat}
}
\description{
Build and execute task graph to slice each partition from the array in
parallel and apply the supplied function.
}
