% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiledbAssay.R
\name{TiledbAssay}
\alias{TiledbAssay}
\title{TileDB Assay}
\description{
TileDB Assay

TileDB Assay
}
\details{
An object to store and manipulate feature/observation matrices in TileDB.
}
\section{Super class}{
\code{\link[tiledbsc:TiledbBase]{tiledbsc::TiledbBase}} -> \code{TiledbAssay}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{URI of the TileDB array}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TiledbAssay$new()}}
\item \href{#method-to_dataframe}{\code{TiledbAssay$to_dataframe()}}
\item \href{#method-to_matrix}{\code{TiledbAssay$to_matrix()}}
\item \href{#method-clone}{\code{TiledbAssay$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="add_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-add_metadata}{\code{tiledbsc::TiledbBase$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="attributes">}\href{../../tiledbsc/html/TiledbBase.html#method-attributes}{\code{tiledbsc::TiledbBase$attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="attrnames">}\href{../../tiledbsc/html/TiledbBase.html#method-attrnames}{\code{tiledbsc::TiledbBase$attrnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="dimensions">}\href{../../tiledbsc/html/TiledbBase.html#method-dimensions}{\code{tiledbsc::TiledbBase$dimensions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="dimnames">}\href{../../tiledbsc/html/TiledbBase.html#method-dimnames}{\code{tiledbsc::TiledbBase$dimnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="get_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-get_metadata}{\code{tiledbsc::TiledbBase$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="schema">}\href{../../tiledbsc/html/TiledbBase.html#method-schema}{\code{tiledbsc::TiledbBase$schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="tiledb_array">}\href{../../tiledbsc/html/TiledbBase.html#method-tiledb_array}{\code{tiledbsc::TiledbBase$tiledb_array()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TiledbAssay object. A new array is created if
an \code{file_path} is provided, otherwise an existing array is opened at
the specified URI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbAssay$new(uri, file_path = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB array}

\item{\code{file_path}}{File path for the assay data to ingest. Support is
currently limited 10X HDF5 files}

\item{\code{verbose}}{Print progress updates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Retrieve the assay data from TileDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbAssay$to_dataframe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-to_matrix}{}}}
\subsection{Method \code{to_matrix()}}{
Retrieve the assay data from TileDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbAssay$to_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbAssay$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
