% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiledbBase.R
\name{TiledbBase}
\alias{TiledbBase}
\title{TileDB Array Base Class}
\description{
TileDB Array Base Class

TileDB Array Base Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{The URI of the TileDB array}

\item{\code{verbose}}{Whether to print verbose output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TiledbBase$new()}}
\item \href{#method-tiledb_array}{\code{TiledbBase$tiledb_array()}}
\item \href{#method-get_metadata}{\code{TiledbBase$get_metadata()}}
\item \href{#method-add_metadata}{\code{TiledbBase$add_metadata()}}
\item \href{#method-schema}{\code{TiledbBase$schema()}}
\item \href{#method-dimensions}{\code{TiledbBase$dimensions()}}
\item \href{#method-attributes}{\code{TiledbBase$attributes()}}
\item \href{#method-dimnames}{\code{TiledbBase$dimnames()}}
\item \href{#method-attrnames}{\code{TiledbBase$attrnames()}}
\item \href{#method-clone}{\code{TiledbBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TiledbBase object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$new(uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI for the TileDB array}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tiledb_array"></a>}}
\if{latex}{\out{\hypertarget{method-tiledb_array}{}}}
\subsection{Method \code{tiledb_array()}}{
Return a \code{\link{tiledbarray}} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$tiledb_array(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments to pass to \code{tiledb::tiledb_array()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[tiledb:tiledb_array]{tiledb::tiledb_array}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Retrieve metadata from a TileDB array.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$get_metadata(key = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve.}

\item{\code{prefix}}{Filter metadata using an optional prefix. Ignored if \code{key}
is not NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of metadata values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Add list of metadata to the specified TileDB array.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$add_metadata(metadata, prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Named list of metadata to add.}

\item{\code{prefix}}{Optional prefix to add to the metadata attribute names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-schema"></a>}}
\if{latex}{\out{\hypertarget{method-schema}{}}}
\subsection{Method \code{schema()}}{
Retrieve the array schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$schema()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{\link[tiledb:tiledb_array_schema]{tiledb::tiledb_array_schema}} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dimensions"></a>}}
\if{latex}{\out{\hypertarget{method-dimensions}{}}}
\subsection{Method \code{dimensions()}}{
Retrieve the array dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$dimensions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of \code{\link[tiledb:tiledb_dim]{tiledb::tiledb_dim}} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attributes"></a>}}
\if{latex}{\out{\hypertarget{method-attributes}{}}}
\subsection{Method \code{attributes()}}{
Retrieve the array attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$attributes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of \code{\link[tiledb:tiledb_attr]{tiledb::tiledb_attr}} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dimnames"></a>}}
\if{latex}{\out{\hypertarget{method-dimnames}{}}}
\subsection{Method \code{dimnames()}}{
Retrieve dimension names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$dimnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector with the array's dimension names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attrnames"></a>}}
\if{latex}{\out{\hypertarget{method-attrnames}{}}}
\subsection{Method \code{attrnames()}}{
Retrieve attribute names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$attrnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector with the array's attribute names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
