% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBGroup.R
\name{TileDBGroup}
\alias{TileDBGroup}
\title{Base class for interacting with TileDB groups}
\value{
A named list of arrays.
}
\description{
Base class for interacting with TileDB groups

Base class for interacting with TileDB groups
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{The URI of the TileDB group}

\item{\code{arrays}}{Named list of arrays in the group}

\item{\code{dimension_name}}{Optional name of the dimension shared by all
arrays within the group (typically \code{obs_id} or \code{var_id}).}

\item{\code{verbose}}{Whether to print verbose output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TileDBGroup$new()}}
\item \href{#method-list_objects}{\code{TileDBGroup$list_objects()}}
\item \href{#method-list_object_uris}{\code{TileDBGroup$list_object_uris()}}
\item \href{#method-get_arrays}{\code{TileDBGroup$get_arrays()}}
\item \href{#method-get_metadata}{\code{TileDBGroup$get_metadata()}}
\item \href{#method-add_metadata}{\code{TileDBGroup$add_metadata()}}
\item \href{#method-clone}{\code{TileDBGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TileDBGroup object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$new(uri, dimension_name = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{TileDB array URI}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_objects"></a>}}
\if{latex}{\out{\hypertarget{method-list_objects}{}}}
\subsection{Method \code{list_objects()}}{
List the TileDB objects within the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_objects(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of object to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all object types are listed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with columns \code{URI} and \code{TYPE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_object_uris"></a>}}
\if{latex}{\out{\hypertarget{method-list_object_uris}{}}}
\subsection{Method \code{list_object_uris()}}{
List URIs for TileDB objects within the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_object_uris(type = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of object to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all object types are listed.}

\item{\code{prefix}}{Filter URIs whose basename contain an optional prefix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of object URIs with names corresponding to the
basename of the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_arrays"></a>}}
\if{latex}{\out{\hypertarget{method-get_arrays}{}}}
\subsection{Method \code{get_arrays()}}{
Retrieve arrays within the group that meet the specified
criteria.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_arrays(prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{String prefix to filter the array names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Retrieve metadata from the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_metadata(key = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve.}

\item{\code{prefix}}{Filter metadata using an optional prefix. Ignored if \code{key}
is not NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of metadata values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Add list of metadata to the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$add_metadata(metadata, prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Named list of metadata to add.}

\item{\code{prefix}}{Optional prefix to add to the metadata attribute names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
