% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledbsc-package.R
\docType{package}
\name{tiledbsc-package}
\alias{tiledbsc}
\alias{tiledbsc-package}
\title{tiledbsc: TileDB Single Cell API}
\description{
Store and retrieve single cell data using TileDB and the on-disk format proposed in the Unified Single Cell Data Model and API. Users can import from and export to in-memory formats used by popular toolchains like Seurat and Bioconductor SingleCellExperiment.
}
\details{
The API is designed around R6-based classes corresponding to the
components described in the \href{https://github.com/single-cell-data/matrix-api/blob/main/specification.md}{data model specification}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TileDB-Inc/tiledbsc}
  \item \url{https://github.com/single-cell-data/matrix-api}
  \item Report bugs at \url{https://github.com/TileDB-Inc/tiledbsc/issues}
}

}
\author{
\strong{Maintainer}: Aaron Wolen \email{aaron@tiledb.com} (\href{https://orcid.org/0000-0003-2542-2202}{ORCID})

Authors:
\itemize{
  \item Dirk Eddelbuettel \email{dirk@tiledb.com} (\href{https://orcid.org/0000-0001-6419-907X}{ORCID})
  \item John Kerl \email{john.kerl@tiledb.com}
}

Other contributors:
\itemize{
  \item TileDB, Inc. [copyright holder, funder]
}

}
\keyword{internal}
