% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiledbVisiumDataset.R
\name{TiledbVisiumDataset}
\alias{TiledbVisiumDataset}
\title{TileDB 10X Visium Dataset}
\description{
TileDB 10X Visium Dataset

TileDB 10X Visium Dataset
}
\details{
Creates a TileDB group containing three TileDB arrays: one for the image
data, one for the image positions, and one for the count data.
}
\section{Super class}{
\code{\link[tiledbsc:TiledbVisiumImage]{tiledbsc::TiledbVisiumImage}} -> \code{TiledbVisiumDataset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array_uri}}{URI of the TileDB array}

\item{\code{assay_array}}{Access the TileDB array containing the assay data}

\item{\code{image_array}}{Access the TileDB array containing the image data}

\item{\code{positions_array}}{Access the TileDB array containing the image
positions}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TiledbVisiumDataset$new()}}
\item \href{#method-to_seurat_object}{\code{TiledbVisiumDataset$to_seurat_object()}}
\item \href{#method-clone}{\code{TiledbVisiumDataset$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbVisiumImage" data-id="to_seurat_visium">}\href{../../tiledbsc/html/TiledbVisiumImage.html#method-to_seurat_visium}{\code{tiledbsc::TiledbVisiumImage$to_seurat_visium()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TiledbVisiumDataset object. A new array is created if \code{count_path} and \code{image_path} are provided, otherwise an existing array is opened at the specified URI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumDataset$new(
  array_uri,
  count_path = NULL,
  image_path = NULL,
  scale_factors_path = NULL,
  image_positions_path = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array_uri}}{URI of the TileDB group}

\item{\code{count_path}}{File path for the h5 file to ingest.}

\item{\code{image_path}}{File path for the image to ingest}

\item{\code{scale_factors_path}}{File path for the scale factors}

\item{\code{image_positions_path}}{File path for the image positions}

\item{\code{verbose}}{Print progress updates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat_object"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat_object}{}}}
\subsection{Method \code{to_seurat_object()}}{
Convert to a Seurat object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumDataset$to_seurat_object(
  project = "SeuratProject",
  assay = "Spatial",
  slice = "slice1"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{\code{\link[SeuratObject:Project]{SeuratObject::Project}} name for the \code{Seurat} object}

\item{\code{assay}}{Name of the initial assay}

\item{\code{slice}}{Name for the stored image of the tissue slice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
