% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiledbVisiumImage.R
\name{TiledbVisiumImage}
\alias{TiledbVisiumImage}
\title{TileDB 10X Visium Image}
\description{
TileDB 10X Visium Image

TileDB 10X Visium Image
}
\details{
Creates a TileDB group containing two TileDB arrays: one for the image data,
and one for the image positions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TiledbVisiumImage$new()}}
\item \href{#method-to_seurat_visium}{\code{TiledbVisiumImage$to_seurat_visium()}}
\item \href{#method-clone}{\code{TiledbVisiumImage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TiledbImage object. A new array is created if
an \code{image_path} is provided, otherwise an existing array is opened at
the specified URI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumImage$new(
  array_uri,
  image_path = NULL,
  scale_factors_path = NULL,
  image_positions_path = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image_path}}{File path for the image to ingest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat_visium"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat_visium}{}}}
\subsection{Method \code{to_seurat_visium()}}{
Convert to a Seurat VisiumV1 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumImage$to_seurat_visium(filter_matrix = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumImage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
