% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTiledbSeurat.R
\name{writeTiledbSeurat}
\alias{writeTiledbSeurat}
\title{Write Seurat Object to TileDB}
\usage{
writeTiledbSeurat(x, base_uri)
}
\arguments{
\item{x}{A \code{\link[SeuratObject:Seurat-class]{SeuratObject::Seurat}} object}

\item{base_uri}{The base directory into which the individual TileDB arrays
will be written.}
}
\description{
Write Seurat Object to TileDB
}
\references{
https://mojaveazure.github.io/seurat-disk/articles/h5Seurat-load.html
}
