% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPairwiseMatrixGroup.R
\name{AnnotationPairwiseMatrixGroup}
\alias{AnnotationPairwiseMatrixGroup}
\title{Single-cell Annotation Matrix Group}
\description{
Class for representing a TileDB group containing one or more
\code{\link{AnnotationPairwiseMatrix}} arrays that share a common dimension name.
}
\section{Super classes}{
\code{\link[tiledbsc:TileDBObject]{tiledbsc::TileDBObject}} -> \code{\link[tiledbsc:TileDBGroup]{tiledbsc::TileDBGroup}} -> \code{\link[tiledbsc:AnnotationGroup]{tiledbsc::AnnotationGroup}} -> \code{AnnotationPairwiseMatrixGroup}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_matrix}{\code{AnnotationPairwiseMatrixGroup$add_matrix()}}
\item \href{#method-add_seurat_graph}{\code{AnnotationPairwiseMatrixGroup$add_seurat_graph()}}
\item \href{#method-clone}{\code{AnnotationPairwiseMatrixGroup$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBObject" data-id="class">}\href{../../tiledbsc/html/TileDBObject.html#method-class}{\code{tiledbsc::TileDBObject$class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBObject" data-id="exists">}\href{../../tiledbsc/html/TileDBObject.html#method-exists}{\code{tiledbsc::TileDBObject$exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_member">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_member}{\code{tiledbsc::TileDBGroup$add_member()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_metadata}{\code{tiledbsc::TileDBGroup$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="count_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-count_members}{\code{tiledbsc::TileDBGroup$count_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_member">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_member}{\code{tiledbsc::TileDBGroup$get_member()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_members}{\code{tiledbsc::TileDBGroup$get_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_metadata}{\code{tiledbsc::TileDBGroup$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="group_exists">}\href{../../tiledbsc/html/TileDBGroup.html#method-group_exists}{\code{tiledbsc::TileDBGroup$group_exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_member_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_member_uris}{\code{tiledbsc::TileDBGroup$list_member_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_members}{\code{tiledbsc::TileDBGroup$list_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_object_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_object_uris}{\code{tiledbsc::TileDBGroup$list_object_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_objects">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_objects}{\code{tiledbsc::TileDBGroup$list_objects()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="print">}\href{../../tiledbsc/html/TileDBGroup.html#method-print}{\code{tiledbsc::TileDBGroup$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="tiledb_group">}\href{../../tiledbsc/html/TileDBGroup.html#method-tiledb_group}{\code{tiledbsc::TileDBGroup$tiledb_group()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="AnnotationGroup" data-id="initialize">}\href{../../tiledbsc/html/AnnotationGroup.html#method-initialize}{\code{tiledbsc::AnnotationGroup$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-add_matrix}{}}}
\subsection{Method \code{add_matrix()}}{
Add a new \code{\link{AnnotationPairwiseMatrix}} array to the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$add_matrix(data, name, metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \code{\link{matrix}} of annotation data to ingest. The \code{matrix} rows
must be aligned to the dimension indicated by the group's
\code{dimension_name}.}

\item{\code{name}}{Name of the new pairwise annotation matrix.}

\item{\code{metadata}}{Named list of metadata to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_seurat_graph"></a>}}
\if{latex}{\out{\hypertarget{method-add_seurat_graph}{}}}
\subsection{Method \code{add_seurat_graph()}}{
Convert a \code{\link[SeuratObject:Graph-class]{SeuratObject::Graph}} object to
\code{\link{AnnotationPairwiseMatrix}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$add_seurat_graph(object, technique)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[SeuratObject:Graph-class]{SeuratObject::Graph}} object.}

\item{\code{technique}}{Name of the technique used to generate the graph
(typically, \code{nn} or \code{snn}).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\subsection{On-Disk Format}{

Arrays are named \verb{graph_<technique>}.
}

\subsection{Metadata}{
\itemize{
\item \code{assay_used}: Name of the assay used to generate the graph.
\item \code{graph_technique}: Name of the technique used to generate the graph.
used.
}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPairwiseMatrixGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
