% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayMatrix.R
\name{AssayMatrix}
\alias{AssayMatrix}
\title{Single-cell Assay Matrix}
\description{
Base class for 2D sparse \code{matrix}-like data with string dimensions. An
\code{AssayMatrix} may contain one or more "layers" (i.e., additional measurements
that share the same dimensions and non-empty coordinates.

Used for the \code{X} field of \code{\link{SCGroup}}.
}
\section{Super class}{
\code{\link[tiledbsc:TileDBArray]{tiledbsc::TileDBArray}} -> \code{AssayMatrix}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{URI of the TileDB array}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-from_matrix}{\code{AssayMatrix$from_matrix()}}
\item \href{#method-from_dataframe}{\code{AssayMatrix$from_dataframe()}}
\item \href{#method-to_dataframe}{\code{AssayMatrix$to_dataframe()}}
\item \href{#method-to_matrix}{\code{AssayMatrix$to_matrix()}}
\item \href{#method-clone}{\code{AssayMatrix$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="add_metadata">}\href{../../tiledbsc/html/TileDBArray.html#method-add_metadata}{\code{tiledbsc::TileDBArray$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="array_exists">}\href{../../tiledbsc/html/TileDBArray.html#method-array_exists}{\code{tiledbsc::TileDBArray$array_exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="attributes">}\href{../../tiledbsc/html/TileDBArray.html#method-attributes}{\code{tiledbsc::TileDBArray$attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="attrnames">}\href{../../tiledbsc/html/TileDBArray.html#method-attrnames}{\code{tiledbsc::TileDBArray$attrnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="class">}\href{../../tiledbsc/html/TileDBArray.html#method-class}{\code{tiledbsc::TileDBArray$class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="dimensions">}\href{../../tiledbsc/html/TileDBArray.html#method-dimensions}{\code{tiledbsc::TileDBArray$dimensions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="dimnames">}\href{../../tiledbsc/html/TileDBArray.html#method-dimnames}{\code{tiledbsc::TileDBArray$dimnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="fragment_count">}\href{../../tiledbsc/html/TileDBArray.html#method-fragment_count}{\code{tiledbsc::TileDBArray$fragment_count()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="get_metadata">}\href{../../tiledbsc/html/TileDBArray.html#method-get_metadata}{\code{tiledbsc::TileDBArray$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="initialize">}\href{../../tiledbsc/html/TileDBArray.html#method-initialize}{\code{tiledbsc::TileDBArray$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="print">}\href{../../tiledbsc/html/TileDBArray.html#method-print}{\code{tiledbsc::TileDBArray$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="schema">}\href{../../tiledbsc/html/TileDBArray.html#method-schema}{\code{tiledbsc::TileDBArray$schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBArray" data-id="tiledb_array">}\href{../../tiledbsc/html/TileDBArray.html#method-tiledb_array}{\code{tiledbsc::TileDBArray$tiledb_array()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-from_matrix}{}}}
\subsection{Method \code{from_matrix()}}{
Ingest assay data from a sparse matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayMatrix$from_matrix(x, index_cols, value_col = "value")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a \code{\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}} or \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}
with string dimensions}

\item{\code{index_cols}}{Names to use for the TileDB array's dimensions that
will contain the matrix row/column names.}

\item{\code{value_col}}{Name to use for the TileDB array's attribute that will
contain the matrix values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-from_dataframe}{}}}
\subsection{Method \code{from_dataframe()}}{
Ingest assay data from a COO-formatted data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayMatrix$from_dataframe(x, index_cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a \code{\link{data.frame}}}

\item{\code{index_cols}}{A column index, either numeric with a column index, or
character with a column name, designating one or more index columns. All
other columns are ingested as attributes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Retrieve the assay data from TileDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayMatrix$to_dataframe(attrs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attrs}}{Specify one or more attributes to retrieve. If \code{NULL},
all attributes are retrieved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-to_matrix}{}}}
\subsection{Method \code{to_matrix()}}{
Retrieve assay data from TileDB as a 2D sparse matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayMatrix$to_matrix(attr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attr}}{The name of the attribute layer to retrieve. If \code{NULL}, the
first layer is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}.
more information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
