% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDataset.R
\name{SCDataset}
\alias{SCDataset}
\title{Single-cell Dataset}
\description{
Class for representing a sc_dataset, which may contain of one or more
\code{\link{SCGroup}}s.
}
\section{Super class}{
\code{\link[tiledbsc:TileDBGroup]{tiledbsc::TileDBGroup}} -> \code{SCDataset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scgroups}}{Named list of \code{\link{SCGroup}}s in the dataset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SCDataset$new()}}
\item \href{#method-from_seurat}{\code{SCDataset$from_seurat()}}
\item \href{#method-to_seurat}{\code{SCDataset$to_seurat()}}
\item \href{#method-scgroup_uris}{\code{SCDataset$scgroup_uris()}}
\item \href{#method-clone}{\code{SCDataset$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_metadata}{\code{tiledbsc::TileDBGroup$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_arrays">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_arrays}{\code{tiledbsc::TileDBGroup$get_arrays()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_metadata}{\code{tiledbsc::TileDBGroup$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_object_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_object_uris}{\code{tiledbsc::TileDBGroup$list_object_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_objects">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_objects}{\code{tiledbsc::TileDBGroup$list_objects()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCDataset object. The existing array group is
opened at the specified array \code{uri} if one is present, otherwise a new
array group is created. The \code{scgroups} field is populated with
\code{SCGroup} objects for each URI passed explicitly to \code{scgroup_uris}, as
well \code{SCGroup} objects discovered within the \code{SCdataset} object's
TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCDataset$new(uri, scgroup_uris = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB group}

\item{\code{scgroup_uris}}{Optional vector of URIs for existing \code{\link{SCGroup}}s to
add to the dataset}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_seurat"></a>}}
\if{latex}{\out{\hypertarget{method-from_seurat}{}}}
\subsection{Method \code{from_seurat()}}{
Convert a Seurat object to a TileDB-backed \code{sc_dataset}.
\subsection{Assays}{

Each \verb{[SeuratObject::Assay}] is converted to a \code{\link{SCGroup}} and written to
a nested TileDB group with a URI of \verb{./scgroup_<assay>} where \verb{<assay>}
is the name of the Seurat assay.
}

\subsection{Dimensionality Reductions}{

Dimensionality reduction results are stored as \code{obsm} and \code{varm} arrays
within an \code{SCGroup}. The \code{\link[SeuratObject:DimReduc-class]{SeuratObject::DimReduc}} object's \code{key} slot
is used to determine which \code{SCGroup} to store the results in. The array
names are \verb{(obsm|varm)_dimreduction_<name>}, where \verb{<name>} is the name
of the dimensionality reduction method (e.g., \code{"pca"}).
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCDataset$from_seurat(object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[SeuratObject:Seurat-class]{SeuratObject::Seurat}} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat}{}}}
\subsection{Method \code{to_seurat()}}{
Convert to a \link[SeuratObject:Seurat-class]{SeuratObject::Seurat} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCDataset$to_seurat(project = "SeuratProject")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{\code{\link[SeuratObject:Project]{SeuratObject::Project}} name for the \code{Seurat} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scgroup_uris"></a>}}
\if{latex}{\out{\hypertarget{method-scgroup_uris}{}}}
\subsection{Method \code{scgroup_uris()}}{
List the \code{\link{SCGroup}} URIs in the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCDataset$scgroup_uris()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of URIs for each \code{\link{SCGroup}} in the dataset.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
