% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGroup.R
\name{SCGroup}
\alias{SCGroup}
\title{Single-cell Group}
\description{
Class for representing a group of TileDB arrays that consitute an \code{sc_group},
which includes:
\itemize{
\item \code{X} (\code{\link{SCGroup_X}}): a labeled 2D sparse array
\item \code{obs} (\code{\link{SCGroup_Annotation}}): 1D labeled array with column labels for \code{X}
\item \code{var} (\code{\link{SCGroup_Annotation}}): 1D labeled array with row labels for \code{X}
}
}
\section{Super class}{
\code{\link[tiledbsc:TiledbGroup]{tiledbsc::TiledbGroup}} -> \code{SCGroup}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obs}}{\code{\link{SCGroup_Annotation}} object containing observation annotations}

\item{\code{var}}{\code{\link{SCGroup_Annotation}} object containing variable annotations}

\item{\code{X}}{\code{\link{SCGroup_X}} object containing assay data}

\item{\code{obsm}}{named list of \code{\link{AnnotationMatrix}} objects aligned with \code{obs}}

\item{\code{varm}}{named list of \code{\link{AnnotationMatrix}} objects aligned with \code{var}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SCGroup$new()}}
\item \href{#method-from_seurat_assay}{\code{SCGroup$from_seurat_assay()}}
\item \href{#method-to_seurat_assay}{\code{SCGroup$to_seurat_assay()}}
\item \href{#method-from_seurat_dimreduction}{\code{SCGroup$from_seurat_dimreduction()}}
\item \href{#method-to_seurat_dimreduction}{\code{SCGroup$to_seurat_dimreduction()}}
\item \href{#method-to_seurat_object}{\code{SCGroup$to_seurat_object()}}
\item \href{#method-clone}{\code{SCGroup$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbGroup" data-id="add_metadata">}\href{../../tiledbsc/html/TiledbGroup.html#method-add_metadata}{\code{tiledbsc::TiledbGroup$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbGroup" data-id="get_metadata">}\href{../../tiledbsc/html/TiledbGroup.html#method-get_metadata}{\code{tiledbsc::TiledbGroup$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbGroup" data-id="list_object_uris">}\href{../../tiledbsc/html/TiledbGroup.html#method-list_object_uris}{\code{tiledbsc::TiledbGroup$list_object_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbGroup" data-id="list_objects">}\href{../../tiledbsc/html/TiledbGroup.html#method-list_objects}{\code{tiledbsc::TiledbGroup$list_objects()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCGroup object. The existing array group is
opened at the specified array \code{uri} if one is present, otherwise a new
array group is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$new(uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB group}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_seurat_assay"></a>}}
\if{latex}{\out{\hypertarget{method-from_seurat_assay}{}}}
\subsection{Method \code{from_seurat_assay()}}{
Convert a Seurat Assay to a TileDB-backed sc_group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$from_seurat_assay(object, obs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[SeuratObject:Assay-class]{SeuratObject::Assay}} object}

\item{\code{obs}}{An optional \code{data.frame} containing annotations for
cell/sample-level observations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat_assay"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat_assay}{}}}
\subsection{Method \code{to_seurat_assay()}}{
Convert to a \code{\link[SeuratObject:Assay-class]{SeuratObject::Assay}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$to_seurat_assay(
  min_cells = 0,
  min_features = 0,
  check_matrix = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_cells}}{Include features detected in at least this many cells.
Will subset the counts matrix as well. To reintroduce excluded features,
create a new object with a lower cutoff.}

\item{\code{min_features}}{Include cells where at least this many features are
detected.}

\item{\code{check_matrix}}{Check counts matrix for NA, NaN, Inf, and non-integer
values}

\item{\code{...}}{Arguments passed to \code{\link[SeuratObject:as.sparse]{SeuratObject::as.sparse}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_seurat_dimreduction"></a>}}
\if{latex}{\out{\hypertarget{method-from_seurat_dimreduction}{}}}
\subsection{Method \code{from_seurat_dimreduction()}}{
Convert a \code{\link[SeuratObject:DimReduc-class]{SeuratObject::DimReduc}} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$from_seurat_dimreduction(object, technique = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[SeuratObject:DimReduc-class]{SeuratObject::DimReduc}} object}

\item{\code{technique}}{Name of the dimensional reduction technique. By default,
the \code{key} slot is used to determine the technique.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\subsection{On-Disk Format}{

Seurat \code{\link{DimReduc}} objects contain a variety of slots to accommodate the
various types of results produced by each of the supported dimensional
reduction techniques. Each slot is stored as an \code{\link{AnnotationMatrix}}
object in the \code{obsm} or \code{varm} slot group for the assay, depending
whether the data is observation- or variable-aligned. The individual
arrays are named \verb{dimreduction_<technique>}.
}

\subsection{Metadata}{
\itemize{
\item \code{dimreduction_technique}: Name of the dimensional reduction technique
used.
\item \code{dimreduction_key}: String prefix used in the dimensional reduction
results column names (required by Seurat)
}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat_dimreduction"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat_dimreduction}{}}}
\subsection{Method \code{to_seurat_dimreduction()}}{
Convert to a \code{\link[SeuratObject:DimReduc-class]{SeuratObject::DimReduc}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$to_seurat_dimreduction(technique = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technique}}{Name of the dimensionality reduction technique. Used to
identify which \code{obsm}/\code{varm} array will be retrieved. If \code{NULL}, we
default to the first \code{obsm/dimreduction_} array.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat_object"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat_object}{}}}
\subsection{Method \code{to_seurat_object()}}{
Convert to a \link[SeuratObject:Seurat-class]{SeuratObject::Seurat} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$to_seurat_object(project = "SeuratProject")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{\code{\link[SeuratObject:Project]{SeuratObject::Project}} name for the \code{Seurat} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
