% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGroup.R
\name{SCGroup}
\alias{SCGroup}
\title{Single-cell Group}
\description{
Class for representing a group of TileDB arrays that consitute an \code{sc_group}, which includes:
\itemize{
\item \code{X} (\code{\link{SCGroup_X}}): a labeled 2D sparse array
\item \code{obs} (\code{\link{SCGroup_Annotation}}): 1D labeled array containing column labels for \code{X}
\item \code{var} (\code{\link{SCGroup_Annotation}}): 1D labeled array containing row labels for \code{X}
}
}
\section{Super class}{
\code{\link[tiledbsc:TiledbGroup]{tiledbsc::TiledbGroup}} -> \code{SCGroup}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obs}}{\code{\link{SCGroup_Annotation}} object containing observation annotations}

\item{\code{var}}{\code{\link{SCGroup_Annotation}} object containing variable annotations}

\item{\code{X}}{\code{\link{SCGroup_X}} object containing assay data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SCGroup$new()}}
\item \href{#method-from_seurat}{\code{SCGroup$from_seurat()}}
\item \href{#method-to_seurat_assay}{\code{SCGroup$to_seurat_assay()}}
\item \href{#method-to_seurat_object}{\code{SCGroup$to_seurat_object()}}
\item \href{#method-clone}{\code{SCGroup$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbGroup" data-id="list_objects">}\href{../../tiledbsc/html/TiledbGroup.html#method-list_objects}{\code{tiledbsc::TiledbGroup$list_objects()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCGoup object. The existing array group is
opened at the specified \code{array_uri} if one is present, otherwise a new
array group is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$new(uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB group}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_seurat"></a>}}
\if{latex}{\out{\hypertarget{method-from_seurat}{}}}
\subsection{Method \code{from_seurat()}}{
Convert a Seurat object to a TileDB-backed sc_group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$from_seurat(object, assay = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{\link[SeuratObject:Seurat-class]{SeuratObject::Seurat}} object.}

\item{\code{assay}}{Name of the assay to retrieve from the Seurat object. By
default the active assay is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat_assay"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat_assay}{}}}
\subsection{Method \code{to_seurat_assay()}}{
Convert to a \code{\link[SeuratObject:Assay-class]{SeuratObject::Assay}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$to_seurat_assay(
  min_cells = 0,
  min_features = 0,
  check_matrix = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_cells}}{Include features detected in at least this many cells.
Will subset the counts matrix as well. To reintroduce excluded features,
create a new object with a lower cutoff.}

\item{\code{min_features}}{Include cells where at least this many features are
detected.}

\item{\code{check_matrix}}{Check counts matrix for NA, NaN, Inf, and non-integer
values}

\item{\code{...}}{Arguments passed to \code{\link[SeuratObject:as.sparse]{SeuratObject::as.sparse}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat_object"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat_object}{}}}
\subsection{Method \code{to_seurat_object()}}{
Convert to a \link[SeuratObject:Seurat-class]{SeuratObject::Seurat} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$to_seurat_object(project = "SeuratProject")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{\code{\link[SeuratObject:Project]{SeuratObject::Project}} name for the \code{Seurat} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
