% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMA.R
\name{SCGroup}
\alias{SCGroup}
\title{Single-cell Group}
\description{
Class for representing the now-deprecated SCGroup object, which has been
renamed to \code{\link{SOMA}}.
}
\section{Super classes}{
\code{\link[tiledbsc:TileDBGroup]{tiledbsc::TileDBGroup}} -> \code{\link[tiledbsc:SOMA]{tiledbsc::SOMA}} -> \code{SCGroup}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{misc}}{An alias for \code{uns}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SCGroup$new()}}
\item \href{#method-clone}{\code{SCGroup$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_member">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_member}{\code{tiledbsc::TileDBGroup$add_member()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="add_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-add_metadata}{\code{tiledbsc::TileDBGroup$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="class">}\href{../../tiledbsc/html/TileDBGroup.html#method-class}{\code{tiledbsc::TileDBGroup$class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="count_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-count_members}{\code{tiledbsc::TileDBGroup$count_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_member">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_member}{\code{tiledbsc::TileDBGroup$get_member()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_members}{\code{tiledbsc::TileDBGroup$get_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="get_metadata">}\href{../../tiledbsc/html/TileDBGroup.html#method-get_metadata}{\code{tiledbsc::TileDBGroup$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="group_exists">}\href{../../tiledbsc/html/TileDBGroup.html#method-group_exists}{\code{tiledbsc::TileDBGroup$group_exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_member_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_member_uris}{\code{tiledbsc::TileDBGroup$list_member_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_members">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_members}{\code{tiledbsc::TileDBGroup$list_members()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_object_uris">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_object_uris}{\code{tiledbsc::TileDBGroup$list_object_uris()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="list_objects">}\href{../../tiledbsc/html/TileDBGroup.html#method-list_objects}{\code{tiledbsc::TileDBGroup$list_objects()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="print">}\href{../../tiledbsc/html/TileDBGroup.html#method-print}{\code{tiledbsc::TileDBGroup$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TileDBGroup" data-id="tiledb_group">}\href{../../tiledbsc/html/TileDBGroup.html#method-tiledb_group}{\code{tiledbsc::TileDBGroup$tiledb_group()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="add_seurat_dimreduction">}\href{../../tiledbsc/html/SOMA.html#method-add_seurat_dimreduction}{\code{tiledbsc::SOMA$add_seurat_dimreduction()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="from_seurat_assay">}\href{../../tiledbsc/html/SOMA.html#method-from_seurat_assay}{\code{tiledbsc::SOMA$from_seurat_assay()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="get_annotation_matrix_arrays">}\href{../../tiledbsc/html/SOMA.html#method-get_annotation_matrix_arrays}{\code{tiledbsc::SOMA$get_annotation_matrix_arrays()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="get_annotation_pairwise_matrix_arrays">}\href{../../tiledbsc/html/SOMA.html#method-get_annotation_pairwise_matrix_arrays}{\code{tiledbsc::SOMA$get_annotation_pairwise_matrix_arrays()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="get_seurat_dimreduction">}\href{../../tiledbsc/html/SOMA.html#method-get_seurat_dimreduction}{\code{tiledbsc::SOMA$get_seurat_dimreduction()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="get_seurat_dimreductions_list">}\href{../../tiledbsc/html/SOMA.html#method-get_seurat_dimreductions_list}{\code{tiledbsc::SOMA$get_seurat_dimreductions_list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="to_seurat_assay">}\href{../../tiledbsc/html/SOMA.html#method-to_seurat_assay}{\code{tiledbsc::SOMA$to_seurat_assay()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="to_seurat_object">}\href{../../tiledbsc/html/SOMA.html#method-to_seurat_object}{\code{tiledbsc::SOMA$to_seurat_object()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="to_single_cell_experiment">}\href{../../tiledbsc/html/SOMA.html#method-to_single_cell_experiment}{\code{tiledbsc::SOMA$to_single_cell_experiment()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="SOMA" data-id="to_summarized_experiment">}\href{../../tiledbsc/html/SOMA.html#method-to_summarized_experiment}{\code{tiledbsc::SOMA$to_summarized_experiment()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCGroup.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$new(uri, verbose = TRUE, config = NULL, ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB group}

\item{\code{verbose}}{Print status messages}

\item{\code{config}}{optional configuration}

\item{\code{ctx}}{optional tiledb context}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
