% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGroup_X.R
\name{SCGroup_X}
\alias{SCGroup_X}
\title{Single-cell Group: X Matrix}
\description{
The \code{X} matrix component of an \code{\link{SCGroup}}.

A \code{SCGroup_X} object is a 2D sparse array with string dimensions \code{obs_id}
and \code{var_id} that map to the dimensions of the \code{obs} and \code{var} arrays,
respectively. An \code{X} matrix can contain or more "layers", which that all
share the same labels/coordinates and non-empty cells.
}
\section{Super class}{
\code{\link[tiledbsc:TiledbBase]{tiledbsc::TiledbBase}} -> \code{SCGroup_X}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array_uri}}{URI of the TileDB array}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SCGroup_X$new()}}
\item \href{#method-from_matrix}{\code{SCGroup_X$from_matrix()}}
\item \href{#method-from_dataframe}{\code{SCGroup_X$from_dataframe()}}
\item \href{#method-to_dataframe}{\code{SCGroup_X$to_dataframe()}}
\item \href{#method-to_matrix}{\code{SCGroup_X$to_matrix()}}
\item \href{#method-clone}{\code{SCGroup_X$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="add_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-add_metadata}{\code{tiledbsc::TiledbBase$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="attributes">}\href{../../tiledbsc/html/TiledbBase.html#method-attributes}{\code{tiledbsc::TiledbBase$attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="attrnames">}\href{../../tiledbsc/html/TiledbBase.html#method-attrnames}{\code{tiledbsc::TiledbBase$attrnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="dimensions">}\href{../../tiledbsc/html/TiledbBase.html#method-dimensions}{\code{tiledbsc::TiledbBase$dimensions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="dimnames">}\href{../../tiledbsc/html/TiledbBase.html#method-dimnames}{\code{tiledbsc::TiledbBase$dimnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="get_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-get_metadata}{\code{tiledbsc::TiledbBase$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="schema">}\href{../../tiledbsc/html/TiledbBase.html#method-schema}{\code{tiledbsc::TiledbBase$schema()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="tiledb_array">}\href{../../tiledbsc/html/TiledbBase.html#method-tiledb_array}{\code{tiledbsc::TiledbBase$tiledb_array()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCObject_X object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$new(array_uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array_uri}}{URI of the TileDB array}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-from_matrix}{}}}
\subsection{Method \code{from_matrix()}}{
Ingest assay data from a sparse matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$from_matrix(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a \code{\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}} or \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}
with string dimensions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-from_dataframe}{}}}
\subsection{Method \code{from_dataframe()}}{
Ingest assay data from a COO-formatted data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$from_dataframe(x, index_cols = c("feature", "barcode"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a \code{\link{data.frame}}}

\item{\code{index_cols}}{A column index, either numeric with a column index, or
character with a column name, designating one or more index columns. All
other columns are ingested as attributes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Retrieve the assay data from TileDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$to_dataframe(attrs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attrs}}{Specify one or more layer attributes to retrieve. If \code{NULL},
all attributes are retrieved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-to_matrix}{}}}
\subsection{Method \code{to_matrix()}}{
Retrieve assay data from TileDB as a 2D sparse matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$to_matrix(attr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attr}}{The name of the attribute layer to retrieve. If \code{NULL}, the
first layer is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[Matrix:dgTMatrix-class]{Matrix::dgTMatrix}}.
more information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCGroup_X$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
