% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiledbGroup.R
\name{TiledbGroup}
\alias{TiledbGroup}
\title{Base class for interacting with TileDB groups}
\description{
Base class for interacting with TileDB groups

Base class for interacting with TileDB groups
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{The URI of the TileDB group}

\item{\code{verbose}}{Whether to print verbose output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TiledbGroup$new()}}
\item \href{#method-list_objects}{\code{TiledbGroup$list_objects()}}
\item \href{#method-get_metadata}{\code{TiledbGroup$get_metadata()}}
\item \href{#method-add_metadata}{\code{TiledbGroup$add_metadata()}}
\item \href{#method-clone}{\code{TiledbGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TiledbGroup object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbGroup$new(uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{TileDB array URI}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_objects"></a>}}
\if{latex}{\out{\hypertarget{method-list_objects}{}}}
\subsection{Method \code{list_objects()}}{
List the TileDB objects within the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbGroup$list_objects(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of object to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all object types are listed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with columns \code{URI} and \code{TYPE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Retrieve metadata from the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbGroup$get_metadata(key = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve.}

\item{\code{prefix}}{Filter metadata using an optional prefix. Ignored if \code{key}
is not NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of metadata values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Add list of metadata to the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbGroup$add_metadata(metadata, prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Named list of metadata to add.}

\item{\code{prefix}}{Optional prefix to add to the metadata attribute names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
