% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiledbGroup.R
\name{TiledbGroup}
\alias{TiledbGroup}
\title{Base class for interacting with TileDB groups}
\description{
Base class for interacting with TileDB groups

Base class for interacting with TileDB groups
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{The URI of the TileDB group}

\item{\code{verbose}}{Whether to print verbose output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TiledbGroup$new()}}
\item \href{#method-list_objects}{\code{TiledbGroup$list_objects()}}
\item \href{#method-clone}{\code{TiledbGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TiledbGroup object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbGroup$new(uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{TileDB array URI}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_objects"></a>}}
\if{latex}{\out{\hypertarget{method-list_objects}{}}}
\subsection{Method \code{list_objects()}}{
List the TileDB objects within the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbGroup$list_objects()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} with columns \code{URI} and \code{TYPE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
