% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBGroup.R
\name{TileDBGroup}
\alias{TileDBGroup}
\title{Base class for interacting with TileDB groups}
\value{
A named list of group members.

The array object.
}
\description{
Base class for interacting with TileDB groups

Base class for interacting with TileDB groups
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{The URI of the TileDB group}

\item{\code{members}}{Named list of members in the group}

\item{\code{verbose}}{Whether to print verbose output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TileDBGroup$new()}}
\item \href{#method-class}{\code{TileDBGroup$class()}}
\item \href{#method-print}{\code{TileDBGroup$print()}}
\item \href{#method-group_exists}{\code{TileDBGroup$group_exists()}}
\item \href{#method-tiledb_group}{\code{TileDBGroup$tiledb_group()}}
\item \href{#method-list_objects}{\code{TileDBGroup$list_objects()}}
\item \href{#method-list_object_uris}{\code{TileDBGroup$list_object_uris()}}
\item \href{#method-add_member}{\code{TileDBGroup$add_member()}}
\item \href{#method-count_members}{\code{TileDBGroup$count_members()}}
\item \href{#method-list_members}{\code{TileDBGroup$list_members()}}
\item \href{#method-list_member_uris}{\code{TileDBGroup$list_member_uris()}}
\item \href{#method-get_members}{\code{TileDBGroup$get_members()}}
\item \href{#method-get_member}{\code{TileDBGroup$get_member()}}
\item \href{#method-get_metadata}{\code{TileDBGroup$get_metadata()}}
\item \href{#method-add_metadata}{\code{TileDBGroup$add_metadata()}}
\item \href{#method-clone}{\code{TileDBGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TileDBGroup object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$new(uri, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{TileDB array URI}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-class"></a>}}
\if{latex}{\out{\hypertarget{method-class}{}}}
\subsection{Method \code{class()}}{
Print the name of the R6 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print summary of the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-group_exists"></a>}}
\if{latex}{\out{\hypertarget{method-group_exists}{}}}
\subsection{Method \code{group_exists()}}{
Check if the group exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$group_exists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if the group exists, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tiledb_group"></a>}}
\if{latex}{\out{\hypertarget{method-tiledb_group}{}}}
\subsection{Method \code{tiledb_group()}}{
Return a \code{\link{tiledb_group}} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$tiledb_group(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments to pass to \code{tiledb::tiledb_array()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[tiledb:tiledb_group]{tiledb::tiledb_group}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_objects"></a>}}
\if{latex}{\out{\hypertarget{method-list_objects}{}}}
\subsection{Method \code{list_objects()}}{
List the TileDB objects within the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_objects(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of object to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all object types are listed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with columns \code{URI} and \code{TYPE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_object_uris"></a>}}
\if{latex}{\out{\hypertarget{method-list_object_uris}{}}}
\subsection{Method \code{list_object_uris()}}{
List URIs for TileDB objects within the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_object_uris(type = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of object to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all object types are listed.}

\item{\code{prefix}}{Filter URIs whose basename contain an optional prefix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of object URIs with names corresponding to the
basename of the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_member"></a>}}
\if{latex}{\out{\hypertarget{method-add_member}{}}}
\subsection{Method \code{add_member()}}{
Add new member to the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$add_member(object, name = NULL, relative = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{The \code{TileDBArray} or \code{TileDBGroup} object to add.}

\item{\code{name}}{The name to use for the member. By default the base name of
the object's URI is used.}

\item{\code{relative}}{A logical value indicating whether the new member's URI
is relative to the group's URI.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count_members"></a>}}
\if{latex}{\out{\hypertarget{method-count_members}{}}}
\subsection{Method \code{count_members()}}{
Count the number of members in the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$count_members()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer count of members in the group.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_members"></a>}}
\if{latex}{\out{\hypertarget{method-list_members}{}}}
\subsection{Method \code{list_members()}}{
List the members of the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_members(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of member to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all member types are listed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with columns \code{URI} and \code{TYPE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_member_uris"></a>}}
\if{latex}{\out{\hypertarget{method-list_member_uris}{}}}
\subsection{Method \code{list_member_uris()}}{
List URIs for group members
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$list_member_uris(type = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of member to list, either \code{"ARRAY"}, or \code{"GROUP"}.
By default all member types are listed.}

\item{\code{prefix}}{Filter URIs whose basename contain an optional prefix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of member URIs with names corresponding to the
basename of the URI.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_members"></a>}}
\if{latex}{\out{\hypertarget{method-get_members}{}}}
\subsection{Method \code{get_members()}}{
Retrieve arrays within the group that meet the specified
criteria.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_members(type = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of group members to list, either \code{"ARRAY"}, or
\code{"GROUP"}.}

\item{\code{prefix}}{String prefix to filter the member names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_member"></a>}}
\if{latex}{\out{\hypertarget{method-get_member}{}}}
\subsection{Method \code{get_member()}}{
Retrieve a group member.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_member(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the array to retrieve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Retrieve metadata from the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_metadata(key = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve.}

\item{\code{prefix}}{Filter metadata using an optional prefix. Ignored if \code{key}
is not NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of metadata values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Add list of metadata to the TileDB group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$add_metadata(metadata, prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Named list of metadata to add.}

\item{\code{prefix}}{Optional prefix to add to the metadata attribute names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
