% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiledbImagePositions.R
\name{TiledbImagePositions}
\alias{TiledbImagePositions}
\title{TileDB Image Position Data}
\description{
TileDB Image Position Data

TileDB Image Position Data
}
\section{Super class}{
\code{\link[tiledbsc:TiledbBase]{tiledbsc::TiledbBase}} -> \code{TiledbImagePositions}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TiledbImagePositions$new()}}
\item \href{#method-to_dataframe}{\code{TiledbImagePositions$to_dataframe()}}
\item \href{#method-clone}{\code{TiledbImagePositions$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="add_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-add_metadata}{\code{tiledbsc::TiledbBase$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="get_metadata">}\href{../../tiledbsc/html/TiledbBase.html#method-get_metadata}{\code{tiledbsc::TiledbBase$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tiledbsc" data-topic="TiledbBase" data-id="tiledb_array">}\href{../../tiledbsc/html/TiledbBase.html#method-tiledb_array}{\code{tiledbsc::TiledbBase$tiledb_array()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TiledbImagePositions object. A new array is created if
an \code{image_positions_path} is provided, otherwise an existing array is opened at
the specified URI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbImagePositions$new(
  array_uri,
  image_positions_path = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image_positions_path}}{File path for the image positions table to ingest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Retrieve the image positions data from TileDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbImagePositions$to_dataframe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbImagePositions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
