% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiledbVisiumImage.R
\name{TiledbVisiumImage}
\alias{TiledbVisiumImage}
\title{TileDB 10X Visium Image}
\description{
TileDB 10X Visium Image

TileDB 10X Visium Image
}
\details{
Creates a TileDB group containing two TileDB arrays: one for the image data,
and one for the image positions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{URI of the TileDB array}

\item{\code{image_array}}{URI of the TileDB array image data}

\item{\code{positions_array}}{Access the TileDB array containing the image}

\item{\code{verbose}}{Print status messages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TiledbVisiumImage$new()}}
\item \href{#method-to_seurat_visium}{\code{TiledbVisiumImage$to_seurat_visium()}}
\item \href{#method-clone}{\code{TiledbVisiumImage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TiledbImage object. A new array is created if
an \code{image_path} is provided, otherwise an existing array is opened at
the specified URI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumImage$new(
  uri,
  image_path = NULL,
  scale_factors_path = NULL,
  image_positions_path = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI of the TileDB group}

\item{\code{image_path}}{File path for the image to ingest}

\item{\code{scale_factors_path}}{File path for the scale factors}

\item{\code{image_positions_path}}{File path for the image positions}

\item{\code{verbose}}{Print progress updates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_seurat_visium"></a>}}
\if{latex}{\out{\hypertarget{method-to_seurat_visium}{}}}
\subsection{Method \code{to_seurat_visium()}}{
Convert to a Seurat VisiumV1 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumImage$to_seurat_visium(filter_matrix = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filter_matrix}}{Only keep spots that have been determined to be over
tissue}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TiledbVisiumImage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
